+++ test/unit/commands/task.test.js
@@ -28,7 +28,6 @@ const cancel = proxyquire('../../../lib/cmds/tasks_cmds/cancel', mocks);
 const retry = proxyquire('../../../lib/cmds/tasks_cmds/retry', mocks);
 const list = proxyquire('../../../lib/cmds/tasks_cmds/list', mocks);
 const create = proxyquire('../../../lib/cmds/tasks_cmds/create', mocks);
-const createFoundationTask = proxyquire('../../../lib/cmds/tasks_cmds/create-foundation-task', mocks);
 const createNantomicsTask = proxyquire('../../../lib/cmds/tasks_cmds/create-nantomics-task', mocks);
 const createAshionTask = proxyquire('../../../lib/cmds/tasks_cmds/create-ashion-task', mocks);
 const createNantBulkTask = proxyquire('../../../lib/cmds/tasks_cmds/create-nant-bulk-task', mocks);
@@ -86,74 +85,6 @@ test.serial.cb('The "tasks-create" command should create a task', t => {
 
   yargs.command(create)
     .parse('create');
-});
-
-test.serial.cb('The "create-foundation-xml-import" command should create a foundation ingest task', t => {
-  const res = { data: {} };
-  postStub.onFirstCall().returns(res);
-  callback = () => {
-    t.is(postStub.callCount, 1);
-    t.is(postStub.getCall(0).args[1], '/v1/tasks/system/foundation-xml-import');
-    t.deepEqual(postStub.getCall(0).args[2], {
-      xmlFileId: 'c8ef7300-1373-4e51-8eb9-ff333600f6a5',
-      datasetId: 'db3e09e9-1ecd-4976-aa5e-70ac7ada0cc3',
-      reportFileId: '1234',
-      subjectId: '2a6dc73e-ed30-4387-94c1-0cd661da56d9',
-      sequenceName: 'test3',
-      testType: 'test1',
-      performerId: 'performer1',
-      indexedDate: '1999-01-01T00:00:00.000Z',
-      indexType: 'all',
-      reIngestFile: false,
-      bodySite: 'Colon',
-      bodySiteSystem: 'http://a.fancy.system.com',
-      bodySiteDisplay: 'body site notation',
-      sendFailedTo: 'failed@user.com'
-    });
-    t.is(printSpy.callCount, 1);
-    t.is(printSpy.getCall(0).args[0], res.data);
-    t.end();
-  };
-
-  yargs.command(createFoundationTask)
-    .parse('create-foundation-xml-import db3e09e9-1ecd-4976-aa5e-70ac7ada0cc3 -x c8ef7300-1373-4e51-8eb9-ff333600f6a5 ' +
-      '-r 1234 -s 2a6dc73e-ed30-4387-94c1-0cd661da56d9 -n test3 --test-type test1 --performer-id performer1 ' +
-      '--indexed-date "1999-01-01" --index-type all --body-site "Colon" --body-site-system "http://a.fancy.system.com" ' +
-      '--body-site-display "body site notation" --send-failed-to failed@user.com');
-});
-
-test.serial.cb('The "create-foundation-xml-import" accepts re-ingest-file as an optional boolean flag', t => {
-  const res = { data: {} };
-  postStub.onFirstCall().returns(res);
-  callback = () => {
-    t.is(postStub.callCount, 1);
-    t.is(postStub.getCall(0).args[1], '/v1/tasks/system/foundation-xml-import');
-    t.deepEqual(postStub.getCall(0).args[2], {
-      xmlFileId: 'c8ef7300-1373-4e51-8eb9-ff333600f6a5',
-      datasetId: 'db3e09e9-1ecd-4976-aa5e-70ac7ada0cc3',
-      reportFileId: '1234',
-      subjectId: '2a6dc73e-ed30-4387-94c1-0cd661da56d9',
-      sequenceName: 'test3',
-      testType: 'test1',
-      performerId: 'performer1',
-      indexedDate: '1999-01-01T00:00:00.000Z',
-      indexType: 'all',
-      reIngestFile: true,
-      bodySite: 'Colon',
-      bodySiteSystem: 'http://a.fancy.system.com',
-      bodySiteDisplay: 'body site notation',
-      sendFailedTo: undefined
-    });
-    t.is(printSpy.callCount, 1);
-    t.is(printSpy.getCall(0).args[0], res.data);
-    t.end();
-  };
-
-  yargs.command(createFoundationTask)
-    .parse('create-foundation-xml-import db3e09e9-1ecd-4976-aa5e-70ac7ada0cc3 -x c8ef7300-1373-4e51-8eb9-ff333600f6a5 ' +
-      '-r 1234 -s 2a6dc73e-ed30-4387-94c1-0cd661da56d9 -n test3 --test-type test1 --performer-id performer1 ' +
-      '--indexed-date "1999-01-01" --index-type all --re-ingest-file --body-site "Colon" ' +
-      '--body-site-system "http://a.fancy.system.com"  --body-site-display "body site notation"');
 });
 
 test.serial.cb('The "create-ashion-import" command should create a ashion ingest task', t => {